+++ components/Episode/EpisodeCard.js
@@ -1,0 +1,5 @@
+import { StyledEpisodeCard } from "./Episode.styled";
+
+export default function EpisodeCard({ children }) {
+  return <StyledEpisodeCard>{children}</StyledEpisodeCard>;
+}
