+++ components/EpisodesList/EpisodeItem.js
@@ -1,0 +1,31 @@
+import SVGIcon from "../Icons";
+import {
+  FavButton,
+  StyledEpisodeLink,
+  StyledListItem,
+} from "./EpisodesList.styled";
+
+export default function EpisodeListItem({
+  episodeNumber,
+  title,
+  href,
+  parts,
+  onHandleFavorites,
+  isFaved,
+}) {
+  return (
+    <StyledListItem>
+      <StyledEpisodeLink href={href}>
+        <p>
+          #{episodeNumber} ...{title} {parts ? "(Spezialfolge)" : ""}
+        </p>
+      </StyledEpisodeLink>
+      <FavButton type="button" onClick={onHandleFavorites}>
+        <SVGIcon
+          variant={isFaved ? "favoriteFilled" : "favoriteEmpty"}
+          width="25px"
+        />
+      </FavButton>
+    </StyledListItem>
+  );
+}
