+++ components/EpisodesList/EpisodesList.js
@@ -1,0 +1,5 @@
+import { StyledList } from "./EpisodesList.styled";
+
+export default function EpisodesList({ children }) {
+  return <StyledList>{children}</StyledList>;
+}
