+++ components/Favoring/favoriteCheck.js
