+++ components/Favoring/initialFavorites.js
